+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ TBD
 - Added Entry Heartbeat event label
 - Updated StatusInfo heartbeat to append no more than one 'NOT_CONFIGURED' to history
 - Removed status, status info, and namespaces from Garden model since status is now derived from connections
+- Added expiration_at and root_command_type to Request model
 
 3.29.1
 ------